AU: SplitWriter class for parsing our full update files.
Full updates now include data for two partitions (kernel + rootfs).
This CL adds a new SplitFileWriter class which takes the stream of
data, which is expected to be in our full update format (8 bytes big
endian uint64_t of the first partition size, followed by first
partition data, followed by second partition size). It parses the
size, then writes the data to each of two FileWriter classes.

BUG=None
TEST=Attached unittest

Review URL: http://codereview.chromium.org/1800009

+++ decompressing_file_writer.cc
@@ -27,7 +27,7 @@ namespace chromeos_update_engine {
 //      uLong   reserved;   /* reserved for future use */
 //  } z_stream;
 
-int GzipDecompressingFileWriter::Write(const void* bytes, size_t count) {
+ssize_t GzipDecompressingFileWriter::Write(const void* bytes, size_t count) {
   // Steps:
   // put the data on next_in
   // call inflate until it returns nothing, each time writing what we get
