+++ file_writer.h
@@ -29,7 +29,7 @@ class FileWriter {
 
   // Wrapper around write. Returns bytes written on success or
   // -errno on error.
-  virtual int Write(const void* bytes, size_t count) = 0;
+  virtual ssize_t Write(const void* bytes, size_t count) = 0;
 
   // Wrapper around close. Returns 0 on success or -errno on error.
   virtual int Close() = 0;
@@ -47,7 +47,7 @@ class DirectFileWriter : public FileWriter {
   virtual ~DirectFileWriter() {}
 
   virtual int Open(const char* path, int flags, mode_t mode);
-  virtual int Write(const void* bytes, size_t count);
+  virtual ssize_t Write(const void* bytes, size_t count);
   virtual int Close();
 
   int fd() const { return fd_; }
